Remove some more redundant stuff
