+++ support-frontend/assets/pages/paper-subscription-checkout/helpers/deliveryDays.js
@@ -13,29 +13,18 @@ import {
   WeekendPlus,
 } from 'helpers/productPrice/productOptions';
 
-type Day = 0 | 1 | 2 | 3 | 4 | 5 | 6;
-const milsInADay = 1000 * 60 * 60 * 24;
+import {
+  getNextDaysOfTheWeek,
+  numberOfWeeksWeDeliverTo,
+} from 'helpers/subscriptionsForms/deliveryDays';
 
-const numberOfWeeksWeDeliverTo = 4;
+import type { Day } from 'helpers/subscriptionsForms/deliveryDays';
+
 // The cut off for getting vouchers in two weeks is Wednesday (day #3 in ISO format) at 6 AM GMT
 const voucherExtraDelayCutoffWeekday = 3;
 const voucherExtraDelayCutoffHour = 6;
 const voucherNormalDelayWeeks = 2;
 const voucherExtraDelayWeeks = 3;
-
-const getNextDayOfTheWeek = (today: number, day: Day): Date => {
-  const diff = (7 + (day - new Date(today).getDay())) % 7;
-  return new Date(today + (diff * milsInADay));
-};
-
-const getNextDaysOfTheWeek = (today: number, day: Day, length: number = numberOfWeeksWeDeliverTo): Date[] => {
-  const initial = getNextDayOfTheWeek(today, day);
-  const rt = [initial];
-  for (let i = 1; i <= length; i += 1) {
-    rt.push(new Date(rt[i - 1].getTime() + (7 * milsInADay)));
-  }
-  return rt;
-};
 
 const getDeliveryDayForProduct = (product: ProductOptions): Day => {
   switch (product) {
@@ -72,5 +61,6 @@ const getVoucherDays = (today: number, product: ProductOptions): Date[] => {
 
 const getDeliveryDays = (today: number, product: ProductOptions): Date[] =>
   getNextDaysOfTheWeek(today, getDeliveryDayForProduct(product), numberOfWeeksWeDeliverTo);
+
 
 export { getVoucherDays, getDeliveryDays };
